# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management API
  description: |
    REST API для управления банковскими картами,
    пользователями, запросами на блокировку
    и переводами между собственными картами.
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/transfers:
    post:
      tags:
      - Переводы
      summary: Перевести средства между своими картами
      description: |
        Выполняет перевод между двумя картами,
        принадлежащими текущему авторизованному пользователю.
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: Перевод успешно выполнен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          description: |
            Некорректная сумма или выбрана
            одна и та же карта
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: |
            Одна из карт принадлежит
            другому пользователю
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Одна из карт не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Недостаточно средств, карта заблокирована
            или срок её действия истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки REST API
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2026-07-28T19:34:52.426
        status:
          type: integer
          format: int32
          description: HTTP-статус ошибки
          example: 404
        error:
          type: string
          description: Наименование HTTP-ошибки
          example: Not Found
        message:
          type: string
          description: Сообщение для пользователя
          example: Карта с идентификатором 15 не найдена
        path:
          type: string
          description: Путь HTTP-запроса
          example: /api/cards/15
...
    TransferRequest:
      type: object
      description: Данные для перевода между своими картами
      properties:
        fromCardId:
          type: integer
          format: int64
          description: Идентификатор карты списания
          example: 1
        toCardId:
          type: integer
          format: int64
          description: Идентификатор карты зачисления
          example: 2
        amount:
          type: number
          description: Сумма перевода
          example: 500.0
          minimum: 0.01
      required:
      - amount
      - fromCardId
      - toCardId
    TransferResponse:
      type: object
      description: Результат перевода между картами
      properties:
        fromCardId:
          type: integer
          format: int64
          description: Идентификатор карты списания
          example: 1
        toCardId:
          type: integer
          format: int64
          description: Идентификатор карты зачисления
          example: 2
        amount:
          type: number
          description: Сумма перевода
          example: 500.0
        fromCardBalance:
          type: number
          description: Баланс карты списания после перевода
          example: 1500.0
        toCardBalance:
          type: number
          description: Баланс карты зачисления после перевода
          example: 3000.0
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT access-токен для авторизации
      scheme: bearer
      bearerFormat: JWT